Precision becomes a ladder, and the SAFETY arrows work
The safety arrows called Reprice, which refreshes numbers and nothing
else. They moved the tier and changed the price while leaving the plate
showing the previous tier's art, name and description, so all three
modules looked broken and finishing a research gave no visible sign it had
landed. They rebuild the stage now, because what changes is the picture.

Precision is FP32 to BF16 to FP8 instead of a single FP8 gate. A company
starts at full width — slow, expensive, never surprising — and buys its way
down. FP32 stays ungated: it is the neutral option, and this catalog has
already learned once what putting a neutral option behind a node costs.

That change failed loudly and correctly on the first run. Twenty-nine tests
went red and the stability scan said every seed bankrupted, which is not
difficulty. Two causes: the new node was marked optionalTechnology, so the
scripted baseline skipped it forever, and the blueprint default was still
BF16, so every test that never mentions precision was building a run it was
no longer allowed to start. Mixed precision is not an option that trades
one thing for another, it is a straight upgrade every lab takes, so it is
not optional; and the blueprint default is the ungated rung.

Saves go to v32. Every company that predates this keeps BF16, because their
models were trained at that width and their costs were priced against it.
Dropping them to full width would raise the bill on a decision they never
made.

The flourish on the MODEL tab is half as long.

633 EditMode, 6 PlayMode, scene 107.

+++ Assets/_ScalingLaws/Scripts/Simulation/ModelBlueprint.cs
@@ -37,7 +37,7 @@ namespace ScalingLaws.Simulation
             DatasetSource dataSources,
             ModelType type = ModelType.General,
             string family = null,
-            TrainingPrecision precision = TrainingPrecision.BFloat16,
+            TrainingPrecision precision = TrainingPrecision.Float32,
             ModelShape shape = ModelShape.Balanced,
             DeduplicationPass deduplication = DeduplicationPass.Standard,
             int cutoffMonthsBack = 0,
